infra-web-app/requirements.txt
aws-cdk-lib==2.185.0
constructs>=10.0.0,<11.0.0
boto3
aws_lambda_powertools
simplejson
python-jose>=3.4.0
jsii
cdk-nag
